020 #Refactor models add new one. Add new endpoints to get reports details and info
